Installation - Fix installation problem with WiX EIDCardLibrary - Fix communication issue when submitting report (didn't look at HTTP return code nor tried to use a proxy) Wizard - Add the same level of functionnality to the XP wizard (compared to the vista one) - Create an "elevated wizard" to correct this cosmectic issue when the windows on the task link were put at the bottom - Show now errors when the user doesn't want to do a test on the wizard (do not know now how to cancel the "finish" button) - Fix resource setup (can now edit both english and french at the same time) - add missing french resource values - Clean not used resource id - Allow not to post "good working" report EIDTest - Fix the "online database" test

+++ EIDConfigurationWizard/DebugReport.cpp
@@ -346,7 +346,7 @@ BOOL CreateDebugReport(PTSTR szLogFile)
 				__leave;
 			}
 			// send to the process the log file name
-			WriteFile(hNamedPipe,szLogFile,(_tcslen(szLogFile) +1) * sizeof(TCHAR),&dwWrite,NULL);
+			WriteFile(hNamedPipe,szLogFile,(DWORD) ((_tcslen(szLogFile) +1) * sizeof(TCHAR)),&dwWrite,NULL);
 			Sleep(1000);
 			// do the action ...
 			DoTheActionToBeTraced();
